Cleaned up APIs Fixed accessibility issue
